+++ crates/onboarding/Cargo.toml
@@ -24,9 +24,13 @@ fs.workspace = true
 gpui.workspace = true
 language.workspace = true
 project.workspace = true
+schemars.workspace = true
+serde.workspace = true
 settings.workspace = true
 theme.workspace = true
 ui.workspace = true
+util.workspace = true
 workspace-hack.workspace = true
 workspace.workspace = true
 zed_actions.workspace = true
+zlog.workspace = true
